app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
geoalchemy2==0.14.2
shapely==2.0.3
httpx==0.27.0
python-dotenv==1.0.0
pandas==2.2.1
jinja2==3.1.3
tenacity==8.3.0
redis==5.0.4